Modification USER

+++ src/main/java/fr/eni/tp/projet/dal/UserDAO.java
@@ -3,9 +3,8 @@ package fr.eni.tp.projet.dal;
 import fr.eni.tp.projet.bo.User;
 
 public interface UserDAO {
-    User findById(int id);
+    User findById(long id);
     User findByEmail(String email);
     User findAllUsers();
     User createUser(User user);
-
 }
